server/requirements.txt
grpcio==1.67.1
grpcio-reflection==1.67.1
grpcio-tools==1.67.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
